refactor nav Fixes #78
