+++ src/libs/ui/index.ts
@@ -7,10 +7,10 @@ export { default as Input } from './atom/input/input';
 export { default as Text } from './atom/text/text';
 
 // molecule components
-export { default as NavBar } from './molecule/nav/nav';
+export { default as NavBar } from './organism/reusable/nav/nav';
 export { default as HeroBox } from './organism/pages/home/head/head';
-export { default as SocialNetwork } from './molecule/socials/socials';
-export { default as Maintenance } from './molecule/maintenance/maintenance';
+export { default as SocialNetwork } from './organism/reusable/socials/socials';
+export { default as Maintenance } from './organism/reusable/maintenance/maintenance';
 
 // organism components
 //pages
